feat: add first working prototype
the prototype will find all vue components with an
<htmlvalidate> block and output content as htmlvalidate rules
for each component.

currently there is not many checks and the code expects a
few things to exist for it to work.

the component name/rule name will be determined in the following order;
name property in the component, followed by the filename. if nothing
is found it will use a default NameNotFound name.

rule content is currently only validated that it is a valid JSON.