feat: support writing slot rules
Html-validate supports writing rules for slots which this commit adds
support for.

In this commit the plugin is updated to handle values that are prefixed
with either `:` or `#` and write these as slot rules.

In addition to the new feature the code is refactor to be a lot more
readable and a bit more documentation.